refactor(lowering): split signature member lowering
